fix: resolve ci/cd check failures
- Fix unused imports in Sandbox and ActionBar components
- Fix TypeScript type annotation for WritableStream in webcontainer
- Fix directory traversal loop bug in file system mounting
- Integrate sandbox routes directly into indexer api.js
- Remove duplicate sandbox-routes.js file
- Update indexer Dockerfile to include migrations directory
- Remove unused session functions imports
