+++ indexer/Dockerfile
@@ -8,6 +8,7 @@ WORKDIR /app
 RUN apk add --no-cache tini wget
 COPY --from=build /app/node_modules ./node_modules
 COPY . .
+COPY migrations ./migrations
 USER node
 ENTRYPOINT ["/sbin/tini", "--"]
 CMD ["node", "src/index.js"]
